pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "teenfreelance"
version = "1.0.0"
description = "Фриланс-платформа для школьников и студентов"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "TeenFreelance Team", email = "team@teenfreelance.ru"}
]
keywords = ["freelance", "teenagers", "students", "fastapi", "nextjs"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

dependencies = [
    # Core Framework
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "pydantic==2.5.0",
    "pydantic[email]==2.5.0",
    "pydantic-settings==2.1.0",
    
    # Database
    "sqlalchemy==2.0.23",
    "aiomysql==0.2.0",
    "alembic==1.13.1",
    
    # Authentication & Security
    "python-jose[cryptography]==3.3.0",
    "PyJWT==2.8.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.6",
    "cryptography>=41.0.0",
    
    # HTTP & Networking
    "httpx==0.25.2",
    "aiofiles==23.2.1",
    "websockets==12.0",
    
    # Monitoring & Logging
    "structlog==23.2.0",
    "sentry-sdk[fastapi]==1.38.0",
    
    # Caching
    "redis==5.0.1",
    "aioredis==2.0.1",
    "zstandard==0.25.0",
    
    # Email
    "fastapi-mail==1.4.1",
    
    # File Processing
    "Pillow==10.1.0",
    "python-magic==0.4.27",
    
    # Production
    "gunicorn==21.2.0",
    "psutil==5.9.6",
]

[project.optional-dependencies]
dev = [
    # Development
    "black==23.11.0",
    "isort==5.12.0",
    "flake8==6.1.0",
    "mypy==1.7.1",
    "pre-commit==3.6.0",
]

test = [
    # Testing
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-cov==4.1.0",
    "pytest-mock==3.12.0",
    "httpx==0.25.2",
]

docs = [
    # Documentation
    "mkdocs==1.5.3",
    "mkdocs-material==9.4.8",
    "mkdocstrings[python]==0.24.0",
]

all = [
    "teenfreelance[dev,test,docs]"
]

[project.urls]
Homepage = "https://github.com/teenfreelance/backend"
Documentation = "https://docs.teenfreelance.ru"
Repository = "https://github.com/teenfreelance/backend"
"Bug Tracker" = "https://github.com/teenfreelance/backend/issues"

[project.scripts]
teenfreelance = "src.main:main"
tf-dev = "src.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

# Black configuration
[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | fronted_new
)/
'''

# isort configuration
[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["src"]
known_third_party = ["fastapi", "pydantic", "sqlalchemy"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

# MyPy configuration
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "aiomysql.*",
    "redis.*",
    "structlog.*",
    "sentry_sdk.*",
    "PIL.*",
    "magic.*",
]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# Coverage configuration
[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

# Ruff configuration (alternative to flake8)
[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.isort]
known-first-party = ["src"]

